requirements.txt
asgiref==3.11.0
Django==6.0
django-environ==0.12.0
django-jalali==7.4.0
jalali_core==1.0.0
jdatetime==5.2.0
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
